# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Basic App Using Gprc",
        "version": "0.1.0"
    },
    "paths": {
        "/create-todo/": {
            "post": {
                "tags": [
                    "Todo Service"
                ],
                "summary": "Create Todo",
                "operationId": "create_todo_create_todo__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTodoSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/get-all-todo/": {
            "get": {
                "tags": [
                    "Todo Service"
                ],
                "summary": "Get All Todo",
                "operationId": "get_all_todo_get_all_todo__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/get-todo/{id}/": {
            "get": {
                "tags": [
                    "Todo Service"
                ],
                "summary": "Get Todo",
                "operationId": "get_todo_get_todo__id___get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/update-todo-status/": {
            "patch": {
                "tags": [
                    "Todo Service"
                ],
                "summary": "Update Todo Status",
                "operationId": "update_todo_status_update_todo_status__patch",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTodoRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/delete-todo/{id}/": {
            "delete": {
                "tags": [
                    "Todo Service"
                ],
                "summary": "Delete Todo",
                "operationId": "delete_todo_delete_todo__id___delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateTodoSchema": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "completed": {
                        "type": "boolean",
                        "title": "Completed"
                    },
                    "day": {
                        "type": "integer",
                        "title": "Day"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "completed",
                    "day"
                ],
                "title": "CreateTodoSchema"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "UpdateTodoRequest": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "completed": {
                        "type": "boolean",
                        "title": "Completed"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "completed"
                ],
                "title": "UpdateTodoRequest"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}